Refactor for next major update
In a nutshell: remove all third-party dependencies which should actually
be linked from the Arduino distribution. This allows us to be up-to-date
and avoid potential legal issues.

Moreover, sources are now named with .cpp extension so that editors work
correctly and moved top-level so we can use `bear` to generate a
`compile_commands.json` that finds all dependencies correctly.

This also marks the start of a more object-oriented structure with
separate interfaces and classes where possible in order to ease moving
components in and out. As of now, the serial communication protocol is
finished and a very basic non-controlling controller can be enabled with
`--with-mock-controller` to test the communcation interface.